+++ ZoomLevel/ModConfig.cs
@@ -11,6 +11,7 @@ namespace ZoomLevel
         public KeybindList KeybindListMaxZoomOrUI { get; set; } = KeybindList.Parse("");
         public KeybindList KeybindListMinZoomOrUI { get; set; } = KeybindList.Parse("");
         public KeybindList KeybindListZoomToCurrentMapSize { get; set; } = KeybindList.Parse("");
+        public KeybindList KeybindListPresetZoomAndUIValues { get; set; } = KeybindList.Parse("");
 
         public KeybindList KeybindListMovementCameraUp { get; set; } = KeybindList.Parse("");
         public KeybindList KeybindListMovementCameraDown { get; set; } = KeybindList.Parse("");
@@ -22,6 +23,7 @@ namespace ZoomLevel
         public KeybindList KeybindListToggleHideUIWithCertainZoom { get; set; } = KeybindList.Parse("");
         public KeybindList KeybindListToggleAnyKeyToResetCamera { get; set; } = KeybindList.Parse("");
         public KeybindList KeybindListToggleAutoZoomToCurrentMapSize { get; set; } = KeybindList.Parse("");
+        public KeybindList KeybindListTogglePresetOnLoadSaveFile { get; set; } = KeybindList.Parse("");
 
         public float ZoomOrUILevelIncreaseValue { get; set; } = 0.05f;
         public float ZoomOrUILevelDecreaseValue { get; set; } = -0.05f;
@@ -30,11 +32,14 @@ namespace ZoomLevel
         public float MinZoomOrUIValue { get; set; } = 0.35f;
         public float ZoomLevelThatHidesUI { get; set; } = 0.35f;
         public int CameraMovementSpeedValue { get; set; } = 15;
+        public float PresetZoomLevelValue { get; set; } = 0.86f;
+        public float PresetUIScaleValue { get; set; } = 0.75f;
 
         public bool SuppressControllerButtons { get; set; } = true;
         public bool AutoZoomToCurrentMapSize { get; set; } = false;
         public bool AnyButtonToCenterCamera { get; set; } = true;
         public bool HideUIWithCertainZoom { get; set; } = false;
+        public bool PresetOnLoadSaveFile { get; set; } = false;
         public bool ZoomAndUIControlEverywhere { get; set; } = false;
     }
 }